+++ src/openhti/database.py
@@ -1,8 +1,5 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-
 """
-SPDX-FileCopyrightText: 2024 Gossamer Technologies
+SPDX-FileCopyrightText: 2025 Gossamer Technologies
 SPDX-License-Identifier: BSD-3-Clause
 
 Database initializer.