+++ tests/test_checksum.py
@@ -1,0 +1,60 @@
+from importlib.resources import files
+from sqlite3 import connect
+from sqlite3 import Row
+from unittest import main
+from unittest import TestCase
+from unittest.mock import MagicMock
+from openhti.checksum import get_checksum
+
+
+class TestChecksum(TestCase):
+    def setUp(self):
+        self.expected = (
+            "797388c088731761d77edfd27335aff46ec07bf5ff23c3dd0d4fda8b4bbb43dc"
+        )
+        path = files("openhti").joinpath("schema.sql")
+        self.schema = path.read_text(encoding="utf-8")
+        self.conn = connect(":memory:")
+        self.conn.row_factory = Row
+        self.conn.executescript(self.schema)
+
+    def tearDown(self):
+        self.conn.close()
+
+    def test_get_checksum_initialize(self):
+        """Test that `get_checksum` returns the appropriate hashed value after database initialization."""
+
+        result = get_checksum(self.conn)
+        self.assertEqual(result, self.expected)
+
+    def test_get_checksum_change(self):
+        """Test that `get_checksum` returns a different hashed value after new data is inserted."""
+
+        self.conn.execute(
+            """
+            INSERT INTO phase(name) VALUES
+                ("test value");
+            """
+        )
+        self.conn.commit()
+        result = get_checksum(self.conn)
+        self.assertNotEqual(result, self.expected)
+
+    def test_get_checksum_no_change(self):
+        """Test that `get_checksum` returns the original hashed value after data is inserted and then deleted."""
+
+        self.conn.execute(
+            """
+            INSERT INTO phase(name) VALUES
+                ("test value");
+            """
+        )
+        self.conn.commit()
+        self.conn.execute("DELETE FROM phase WHERE id = 1")
+        self.conn.commit()
+        result = get_checksum(self.conn)
+        self.assertEqual(result, self.expected)
+
+
+if __name__ == "__main__":
+    main()