+++ tests/test_factory.py
@@ -1,6 +1,3 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-
 from unittest import IsolatedAsyncioTestCase
 from unittest import main