+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.0] - 2025-09-25
+
+### Added
+
+- Ability to initialize/reset the database using the UI.
+- Ability to detect dirty/modified database tables using a checksum and manually reset the checksum.
+
 ## [0.0.2] - 2025-04-25
 
 ### Fixed
@@ -24,3 +31,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [0.0.1]: https://github.com/mcpcpc/openhti/releases/tag/0.0.1
 [0.0.2]: https://github.com/mcpcpc/openhti/releases/tag/0.0.2
+[0.1.0]: https://github.com/mcpcpc/openhti/releases/tag/0.1.0